refactor: rename to dashboard-state-factory
