+++ src/data/dashboard-file.ts
@@ -1,6 +1,6 @@
 import { Notice } from "obsidian";
 import { createFile, createFolder } from "./file-operations";
-import { createDashboardState } from "./table-state-factory";
+import { createDashboardState } from "./dashboard-state-factory";
 import { serializeDashboardState } from "./serialize-dashboard-state";
 import { DEFAULT_TABLE_NAME, PREVIOUS_FILE_EXTENSION } from "./constants";
