+++ src/shared/commands/cell-header-update-command.spec.ts
@@ -1,4 +1,4 @@
-import { createDashboardState } from "src/data/table-state-factory";
+import { createDashboardState } from "src/data/dashboard-state-factory";
 import { CommandUndoError } from "./command-errors";
 import CellHeaderUpdateCommand from "./cell-header-update-command";
