+++ src/shared/commands/column-add-command.spec.ts
@@ -1,4 +1,4 @@
-import { createDashboardState } from "src/data/table-state-factory";
+import { createDashboardState } from "src/data/dashboard-state-factory";
 import { CommandRedoError, CommandUndoError } from "./command-errors";
 import ColumnAddCommand from "./column-add-command";
