+++ src/shared/commands/column-type-update-command.ts
@@ -1,4 +1,4 @@
-import { createTag } from "src/data/table-state-factory";
+import { createTag } from "src/data/dashboard-state-factory";
 import { unixTimeToDateString } from "../date/date-conversion";
 import { CHECKBOX_MARKDOWN_UNCHECKED } from "../dashboard-state/constants";
 import { ColumNotFoundError } from "../dashboard-state/dashboard-error";
