+++ src/shared/commands/row-add-command.spec.ts
@@ -1,4 +1,4 @@
-import { createDashboardState } from "src/data/table-state-factory";
+import { createDashboardState } from "src/data/dashboard-state-factory";
 import RowAddCommand from "./row-add-command";
 import { CommandRedoError, CommandUndoError } from "./command-errors";
