+++ src/shared/commands/row-delete-command.spec.ts
@@ -1,4 +1,4 @@
-import { createDashboardState } from "src/data/table-state-factory";
+import { createDashboardState } from "src/data/dashboard-state-factory";
 import RowDeleteCommand from "./row-delete-command";
 import {
 	DeleteCommandArgumentsError,
