+++ src/shared/commands/tag-add-command.ts
@@ -1,4 +1,4 @@
-import { createTag } from "src/data/table-state-factory";
+import { createTag } from "src/data/dashboard-state-factory";
 import DashboardStateCommand from "../dashboard-state/dashboard-state-command";
 import { DashboardState, Tag } from "../types";
 import { Color } from "../types";
