+++ src/shared/commands/tag-cell-remove-command.spec.ts
@@ -1,4 +1,7 @@
-import { createDashboardState, createTag } from "src/data/table-state-factory";
+import {
+	createDashboardState,
+	createTag,
+} from "src/data/dashboard-state-factory";
 import { CommandUndoError } from "./command-errors";
 import TagCellRemoveCommand from "./tag-cell-remove-command";
 import { advanceBy, clear } from "jest-date-mock";
