+++ src/shared/commands/tag-delete-command.spec.ts
@@ -1,4 +1,7 @@
-import { createDashboardState, createTag } from "src/data/table-state-factory";
+import {
+	createDashboardState,
+	createTag,
+} from "src/data/dashboard-state-factory";
 import { CommandUndoError, CommandRedoError } from "./command-errors";
 import TagDeleteCommand from "./tag-delete-command";
